第二次作业的提交

+++ exercises/conversions/using_as.rs
@@ -8,13 +8,11 @@
 // returns the proper type.
 //
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
-// hint.
 
-// I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
